dix: add new lookup functions for input devices via config_string
