+++ config/config-backends.h
@@ -27,6 +27,5 @@
 #define XSERVER_CONFIG_BACKENDS_H
 
 void remove_devices(const char *backend, const char *config_info);
-BOOL device_is_duplicate(const char *config_info);
 
 #endif /* XSERVER_CONFIG_BACKENDS_H */
